Rename  airspeed_options.h to options_airspeed.h for MPLAB-X projects.
