+++ MatrixPilot/helicalTurnCntrl.c
@@ -31,7 +31,7 @@
 #include "../libDCM/mathlibNAV.h"
 #include "../libDCM/rmat.h"
 #include <math.h>
-#include "airspeed_options.h" 
+#include "options_airspeed.h" 
 
 //#ifndef RTL_PITCH_DOWN
 //#define RTL_PITCH_DOWN (0.0)
